+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Perf
 
+### 2025-11-07
+- Reuse stack pool in LEVM [#5179](https://github.com/lambdaclass/ethrex/pull/5179)
+
 ### 2025-11-03
 
 - Avoid unnecessary hash validations [#5167](https://github.com/lambdaclass/ethrex/pull/5167)